+++ nessi/globopt/tests/test_init_pspace.py
@@ -1,5 +1,5 @@
 import numpy as np
-from nessi.pso import Swarm
+from nessi.globopt import Swarm
 
 
 def test_init_pspace_one_line():
@@ -19,7 +19,7 @@ def test_init_pspace_one_line():
 
 def test_init_pspace_multi_line():
     """
-    swarm.init_pspace test for multipoint search.
+    swarm.init_pspace test for multipoints search.
     """
     pspace_output = np.zeros((3, 2, 3), dtype=np.float32)
     pspace_output = np.array([
@@ -50,6 +50,56 @@ def test_init_particles():
     swarm.init_pspace('data/pspace_multi_line.txt')
     swarm.init_particles(2)
     np.testing.assert_equal(swarm.current, output)
-    
+
+def test_get_gbest_full():
+    """
+    swarm.get_best tests.
+    """
+    output = np.array([[-0.4848329 ,  1.111317  ]], dtype=np.float32)
+
+
+    np.random.seed(1)
+    swarm = Swarm()
+    swarm.init_pspace('data/pspace_one_line.txt')
+    swarm.init_particles(9)
+    # Fake evaluation
+    swarm.history[:, :, :] = swarm.current[:, :, :]
+    swarm.misfit[:] = 1.0
+    swarm.misfit[5] = 0.0
+    np.testing.assert_equal(swarm.get_gbest(topology='full'), output)
+
+def test_get_gbest_ring():
+    """
+    swarm.get_best tests.
+    """
+    output = np.array([[-0.4848329 ,  1.111317  ]], dtype=np.float32)
+
+
+    np.random.seed(1)
+    swarm = Swarm()
+    swarm.init_pspace('data/pspace_one_line.txt')
+    swarm.init_particles(9)
+    # Fake evaluation
+    swarm.history[:, :, :] = swarm.current[:, :, :]
+    swarm.misfit[:] = 1.0
+    swarm.misfit[5] = 0.0
+    np.testing.assert_equal(swarm.get_gbest(topology='ring', indv=4), output)
+
+def test_get_gbest_toroidal():
+    """
+    swarm.get_best tests.
+    """
+    output = np.array([[-0.4848329 ,  1.111317  ]], dtype=np.float32)
+
+    np.random.seed(1)
+    swarm = Swarm()
+    swarm.init_pspace('data/pspace_one_line.txt')
+    swarm.init_particles(9)
+    # Fake evaluation
+    swarm.history[:, :, :] = swarm.current[:, :, :]
+    swarm.misfit[:] = 1.0
+    swarm.misfit[5] = 0.0
+    np.testing.assert_equal(swarm.get_gbest(topology='toroidal', indv=8, ndim=3), output)
+
 if __name__ == "__main__" :
     np.testing.run_module_suite()